Add action & itemGroup creators to react-ui-menu.
