+++ packages/ui/react-ui-editor/src/components/EditorToolbar/util.ts
@@ -4,7 +4,6 @@
 
 import { useMemo } from 'react';
 
-import { ACTION_GROUP_TYPE, ACTION_TYPE, actionGroupSymbol } from '@dxos/app-graph';
 import { create, type ReactiveObject } from '@dxos/live-object';
 import { type Label, type ThemedClassName } from '@dxos/react-ui';
 import {
@@ -12,6 +11,8 @@ import {
   type MenuItemGroup,
   type ToolbarMenuActionGroupProperties,
   type MenuActionProperties,
+  createMenuAction,
+  createMenuItemGroup,
 } from '@dxos/react-ui-menu';
 
 import type { EditorAction, EditorActionPayload, EditorViewMode, Formatting } from '../../extensions';
@@ -43,40 +44,18 @@ export type EditorToolbarProps = ThemedClassName<EditorToolbarActionGraphProps &
 
 export type EditorToolbarItem = EditorAction | MenuItemGroup | MenuSeparator;
 
-const noop = () => {};
-
 export const createEditorAction = (
   payload: EditorActionPayload & Partial<MenuActionProperties>,
   icon: string,
   label: Label = [`${payload.type} label`, { ns: translationKey }],
   id: string = payload.type,
-) =>
-  ({
-    id,
-    type: ACTION_TYPE,
-    properties: {
-      ...payload,
-      label,
-      icon,
-    },
-    data: noop,
-  }) satisfies EditorAction;
+) => createMenuAction(id, { icon, label, ...payload }) as EditorAction;
 
 export const createEditorActionGroup = (
   id: string,
-  props: Omit<ToolbarMenuActionGroupProperties, 'label' | 'icon'>,
+  props: Omit<ToolbarMenuActionGroupProperties, 'icon'>,
   icon?: string,
-) =>
-  ({
-    id,
-    type: ACTION_GROUP_TYPE,
-    properties: {
-      ...props,
-      label: [`${id} label`, { ns: translationKey }],
-      icon,
-    } as ToolbarMenuActionGroupProperties,
-    data: actionGroupSymbol,
-  }) satisfies MenuItemGroup;
+) => createMenuItemGroup(id, { icon, ...props });
 
 export const editorToolbarGap = {
   id: 'gap',
