+++ packages/ui/react-ui-menu/src/index.ts
@@ -5,3 +5,4 @@
 export * from './components';
 export * from './defs';
 export * from './hooks';
+export { createMenuAction, createMenuItemGroup } from './util';
