+++ packages/ui/react-ui-menu/src/util.ts
@@ -2,8 +2,10 @@
 // Copyright 2025 DXOS.org
 //
 
-import { type ActionLike } from '@dxos/app-graph';
+import { ACTION_GROUP_TYPE, ACTION_TYPE, actionGroupSymbol, type ActionLike } from '@dxos/app-graph';
 import { getHostPlatform } from '@dxos/util';
+
+import { type MenuAction, type MenuActionProperties, type MenuItemGroup, type MenuItemGroupProperties } from './defs';
 
 export const getShortcut = (action: ActionLike) => {
   return typeof action.properties?.keyBinding === 'string'
@@ -12,3 +14,26 @@ export const getShortcut = (action: ActionLike) => {
 };
 
 export const fallbackIcon = 'ph--placeholder--regular';
+
+const noop = () => {};
+
+export const createMenuAction = <P extends MenuActionProperties = MenuActionProperties>(id: string, properties: P) =>
+  ({
+    id,
+    type: ACTION_TYPE,
+    properties,
+    data: noop,
+  }) satisfies MenuAction;
+
+export const createMenuItemGroup = <
+  P extends MenuItemGroupProperties = MenuItemGroupProperties & Partial<MenuActionProperties>,
+>(
+  id: string,
+  properties: P,
+) =>
+  ({
+    id,
+    type: ACTION_GROUP_TYPE,
+    properties,
+    data: actionGroupSymbol,
+  }) satisfies MenuItemGroup;
